Fix a number of issues that were causing the network to not learn anything
Weights are now properly copied over to the training phase. Weights are no longer only positive, so do not explode the input. Relu is replaced with sigmoid to avoid exploding. The impage input is now correctly stored and used on both sides of WebAssembly as the 0.0 to 1.0 floating point values it always was. It is no longer erroneously treated as an u8 at any stage.

+++ www/worker.js
@@ -31,7 +31,7 @@ wasm_bindgen('pkg/mnist_wasm_bg.wasm').then(mnistWasmModule => {
                 let image = training.images[i]
                 let label = training.labels[i]
                 let imageWasm = Image.new()
-                let pixels = new Uint8Array(memory.buffer, imageWasm.buffer(), WIDTH * HEIGHT)
+                let pixels = new Float64Array(memory.buffer, imageWasm.buffer(), WIDTH * HEIGHT)
                 // copy each pixel into the buffer exposed over Wasm to give it to
                 // the Rust code
                 for (let j = 0; j < WIDTH * HEIGHT; j++) {
